Fix: delete role property from README.md
As far as I understand and being experienced with the API. Proving the role property causes the Meeting SDK to enter a state where it is not possible to join the meeting. The role property is not required and the Meeting SDK will automatically assign the role based on the token provided by the Auth endpoint.

I updated the README.md to reflect this, because it is confusing to have the role property in the example code and not have it work when I was trying to use the Meeting SDK.

Please let me know if the change makes sense.